Upgrade to Wrangler 4.32.0 and simplify configuration
- Upgrade Wrangler from 3.114.14 to 4.32.0
- Remove redundant wrangler.production.toml file
- Consolidate to single wrangler.toml configuration
- Remove environment-specific sections and logic
- Update package.json scripts for simplified deployment
- Add better documentation for local development limitations
- Update test files with clearer guidance about local vs production
- Add .wrangler/tmp/ to .gitignore
- All tests passing - GitHub and Twitch integrations working correctly
