+++ clip-recap-pipeline/.gitignore
@@ -1,0 +1,1 @@
+.wrangler/tmp/
