--- clip-recap-pipeline/wrangler.production.toml
@@ -1,37 +1,0 @@
-name = "clip-recap-pipeline"
-main = "src/index.ts"
-compatibility_date = "2024-01-15"
-
-# Cron triggers for production. All times are UTC.
-[triggers]
-crons = [
-  "0 2 * * *",  # Daily at 02:00 UTC (= 09:00 ICT) - main pipeline
-  "0 * * * *"   # Every hour - token validation
-]
-
-[ai]
-binding = "ai"
-
-[[vectorize]]
-binding = "VECTORIZE"
-index_name = "clip-recap-embeddings"
-
-[[r2_buckets]]
-binding = "R2_BUCKET"
-bucket_name = "clip-recap-assets"
-
-# Environment variables (set via wrangler secret put)
-# TWITCH_CLIENT_ID
-# TWITCH_CLIENT_SECRET  
-# TWITCH_BROADCASTER_ID
-# TWITCH_BROADCASTER_LOGIN
-# GITHUB_APP_ID
-# GITHUB_INSTALLATION_ID
-# GITHUB_PRIVATE_KEY  # PEM-encoded private key for the GitHub App
-# GITHUB_WEBHOOK_SECRET
-# DISCORD_BOT_TOKEN
-# DISCORD_REVIEW_CHANNEL_ID
-# CONTENT_REPO_OWNER
-# CONTENT_REPO_NAME
-# CONTENT_REPO_MAIN_BRANCH
-# CONTENT_REPO_STAGING_BRANCH
